+++ appTools/ToolNCC.py
@@ -3784,7 +3784,8 @@ class NccUI:
         self.addtool_from_db_btn = QtWidgets.QPushButton(_('Add from DB'))
         self.addtool_from_db_btn.setToolTip(
             _("Add a new tool to the Tool Table\n"
-              "from the Tool DataBase.")
+              "from the Tool Database.\n"
+              "Tool database administration in Menu: Options -> Tools Database")
         )
 
         hlay.addWidget(self.addtool_btn)